refactor: transform finance module
