+++ src/index.ts
@@ -58,15 +58,13 @@ export type { CompanyModule } from './modules/company';
 export type { DatabaseModule } from './modules/database';
 export type { DatatypeModule } from './modules/datatype';
 export type { DateModule, SimpleDateModule } from './modules/date';
-export type { Currency, FinanceModule } from './modules/finance';
-export {
-  BitcoinAddressFamily,
-  BitcoinNetwork,
-} from './modules/finance/_bitcoin';
+export { BitcoinAddressFamily, BitcoinNetwork } from './modules/finance';
 export type {
   BitcoinAddressFamilyType,
   BitcoinNetworkType,
-} from './modules/finance/_bitcoin';
+  Currency,
+  FinanceModule,
+} from './modules/finance';
 export type { FoodModule } from './modules/food';
 export type { GitModule } from './modules/git';
 export type { HackerModule } from './modules/hacker';
