+++ src/modules/finance/currency.ts
@@ -2,6 +2,31 @@ import type { FakerCore } from '../../core';
 import { arrayElement } from '../helpers/array-element';
 
 /**
+ * The possible definitions related to currency entries.
+ */
+export interface Currency {
+  /**
+   * The full name for the currency (e.g. `US Dollar`).
+   */
+  name: string;
+
+  /**
+   * The code/short text/abbreviation for the currency (e.g. `USD`).
+   */
+  code: string;
+
+  /**
+   * The symbol for the currency (e.g. `$`).
+   */
+  symbol: string;
+
+  /**
+   * The ISO 4217 numeric code for the currency (e.g. `840`).
+   */
+  numericCode: string;
+}
+
+/**
  * Returns a random currency object, containing `code`, `name`, `symbol`, and `numericCode` properties.
  *
  * @param fakerCore The FakerCore to use.
