+++ src/modules/finance/iban.ts
@@ -3,6 +3,23 @@ import { FakerError } from '../../errors/faker-error';
 import { boolean } from '../datatype/boolean';
 import { arrayElement } from '../helpers/array-element';
 import { int } from '../number/int';
+import { ibanLib } from './_iban-lib';
+
+/**
+ * Puts a space after every 4 characters.
+ *
+ * @internal
+ *
+ * @param iban The iban to pretty print.
+ */
+export function prettyPrintIban(iban: string): string {
+  let pretty = '';
+  for (let i = 0; i < iban.length; i += 4) {
+    pretty += `${iban.substring(i, i + 4)} `;
+  }
+
+  return pretty.trimEnd();
+}
 
 /**
  * Generates a random IBAN.
@@ -44,8 +61,8 @@ export function iban(
   const { countryCode, formatted = false } = options;
 
   const ibanFormat = countryCode
-    ? iban.formats.find((f) => f.country === countryCode)
-    : arrayElement(fakerCore, iban.formats);
+    ? ibanLib.formats.find((f) => f.country === countryCode)
+    : arrayElement(fakerCore, ibanLib.formats);
 
   if (!ibanFormat) {
     throw new FakerError(`Country code ${countryCode} not supported.`);
@@ -58,20 +75,20 @@ export function iban(
     count += bban.count;
     while (c > 0) {
       if (bban.type === 'a') {
-        s += arrayElement(fakerCore, iban.alpha);
+        s += arrayElement(fakerCore, ibanLib.alpha);
       } else if (bban.type === 'c') {
         if (boolean(fakerCore, 0.8)) {
           s += int(fakerCore, 9);
         } else {
-          s += arrayElement(fakerCore, iban.alpha);
+          s += arrayElement(fakerCore, ibanLib.alpha);
         }
       } else {
         if (c >= 3 && boolean(fakerCore, 0.3)) {
           if (boolean(fakerCore)) {
-            s += arrayElement(fakerCore, iban.pattern100);
+            s += arrayElement(fakerCore, ibanLib.pattern100);
             c -= 2;
           } else {
-            s += arrayElement(fakerCore, iban.pattern10);
+            s += arrayElement(fakerCore, ibanLib.pattern10);
             c--;
           }
         } else {
@@ -86,7 +103,7 @@ export function iban(
   }
 
   let checksum: string | number =
-    98 - iban.mod97(iban.toDigitString(`${s}${ibanFormat.country}00`));
+    98 - ibanLib.mod97(ibanLib.toDigitString(`${s}${ibanFormat.country}00`));
 
   if (checksum < 10) {
     checksum = `0${checksum}`;
