+++ src/modules/finance/index.ts
@@ -1,1 +1,7 @@
+export { BitcoinAddressFamily, BitcoinNetwork } from './bitcoin-address';
+export type {
+  BitcoinAddressFamilyType,
+  BitcoinNetworkType,
+} from './bitcoin-address';
+export type { Currency } from './currency';
 export * from './module';
